@@ -4,6 +4,7 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
+    "prebuild": "node scripts/generate-build-info.js",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test --code-coverage",
@@ -11,7 +12,7 @@
     "http-server": "npx http-server -p 8080 -c-1 dist/spiewnigator-angular/browser",
     "build-deploy": "ng build --base-href=/spiewnigator-dev/",
     "lint": "ng lint",
-    "serve-coverage": "http-server -p 9000 -c-1 coverage/spiewnigator-angular"
+    "eslint": "eslint \"src/**/*.{ts,html}\" --cache --ext .ts,.html"
   },
   "private": true,
   "dependencies": {
@@ -26,6 +27,7 @@
     "@angular/platform-browser-dynamic": "^20.3.15",
     "@angular/router": "^20.3.15",
     "@angular/service-worker": "^20.3.15",
+    "fuse.js": "^7.1.0",
     "rxjs": "~6.6.0",
     "tslib": "^2.6.3",
     "zone.js": "^0.15.1"
@@ -36,14 +38,16 @@
     "@angular-eslint/eslint-plugin-template": "20.7.0",
     "@angular-eslint/schematics": "20.7.0",
     "@angular-eslint/template-parser": "20.7.0",
-    "@angular/build": "^20.3.12",
-    "@angular/cli": "^20.3.12",
+    "@angular/build": "^20.3.13",
+    "@angular/cli": "^20.3.13",
     "@angular/compiler-cli": "^20.3.15",
+    "@eslint/eslintrc": "^3.3.1",
+    "@eslint/js": "^9.38.0",
     "@types/jasmine": "~3.8.0",
     "@types/node": "^24.10.1",
-    "@typescript-eslint/eslint-plugin": "^8.38.0",
-    "@typescript-eslint/parser": "^8.38.0",
-    "eslint": "^9.33.0",
+    "@typescript-eslint/eslint-plugin": "^8.46.2",
+    "@typescript-eslint/parser": "^8.46.2",
+    "eslint": "^9.38.0",
     "jasmine-core": "~3.8.0",
     "karma": "^6.4.0",
     "karma-chrome-launcher": "~3.1.0",
