@@ -1,11 +1,33 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { MatToolbar } from '@angular/material/toolbar';
 import { MatAccordion, MatExpansionPanel, MatExpansionPanelHeader, MatExpansionPanelTitle } from '@angular/material/expansion';
+import { Observable, from } from 'rxjs';
+import { DatePipe } from '@angular/common';
 
 @Component({
     selector: 'app-help',
     templateUrl: './help.component.html',
     styleUrls: ['./help.component.scss'],
-    imports: [MatToolbar, MatAccordion, MatExpansionPanel, MatExpansionPanelHeader, MatExpansionPanelTitle]
+    imports: [MatToolbar, MatAccordion, MatExpansionPanel, MatExpansionPanelHeader, MatExpansionPanelTitle, DatePipe]
 })
-export class HelpComponent {}
+export class HelpComponent implements OnInit {
+
+    buildTimestamp?: Date;
+
+    ngOnInit(): void {
+        this.getBuildTimestamp().subscribe(ts => {
+            this.buildTimestamp = ts;
+        });
+    }
+
+    private getBuildTimestamp(): Observable<Date | undefined> {
+        const promise = fetch('/assets/build-info.json')
+            .then(res => res.json())
+            .then(data => new Date(data.timestamp))
+            .catch((e) => {
+                console.warn('Could not load build info', e);
+                return undefined;
+            });
+        return from(promise);
+    }
+}
