@@ -1,4 +1,6 @@
-import { enableProdMode, importProvidersFrom } from '@angular/core';
+import { enableProdMode, importProvidersFrom, LOCALE_ID } from '@angular/core';
+import { registerLocaleData } from '@angular/common';
+import localePl from '@angular/common/locales/pl';
 
 import { environment } from './environments/environment';
 import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
@@ -8,6 +10,8 @@ import { AppRoutingModule } from './app/app-routing.module';
 import { provideAnimations } from '@angular/platform-browser/animations';
 import { ServiceWorkerModule } from '@angular/service-worker';
 import { AppComponent } from './app/app.component';
+
+registerLocaleData(localePl);
 
 if (environment.production) {
   enableProdMode();
@@ -22,7 +26,8 @@ bootstrapApplication(AppComponent, {
             registrationStrategy: 'registerWhenStable:30000'
         })),
         provideHttpClient(withInterceptorsFromDi()),
-        provideAnimations()
+        provideAnimations(),
+        { provide: LOCALE_ID, useValue: 'pl-PL' }
     ]
 })
   .catch(err => console.error(err));
